cleanup code of op3_gui_demo
