+++ op3_gui_demo/config/gui_config.yaml
@@ -20,11 +20,11 @@ id_joint:
   19 : head_pan 
   20 : head_tilt
   200 : opencr
-module_list: [none, head_control_module, base_module, walking_module, manipulation_module, action_module, walking_module, direct_control_module, wholebody_module]
+module_list: [none, head_control_module, base_module, walking_module, manipulation_module, action_module, walking_module, direct_control_module, online_walking_module]
 module_button:
   1 : none
   2 : head_control_module
   3 : action_module
   4 : walking_module
-  5 : wholebody_module
+  5 : online_walking_module
   6 : direct_control_module
