+++ op3_gui_demo/launch/op3_demo.launch
@@ -2,7 +2,7 @@
 
 <launch>
   <arg name="args" default=""/>
-  <param name="demo_config" value="$(find op3_gui_demo)/config/demo_config.yaml"/>
+  <param name="demo_config" value="$(find op3_gui_demo)/config/gui_config.yaml"/>
   
   <node name="op3_demo_opc" pkg="op3_gui_demo" type="op3_gui_demo" output="screen" args="$(arg args)">
     <remap from="/op3_demo/ik_target_pose" to="/pose_panel/pose" />
