+++ op3_gui_demo/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>op3_gui_demo</name>
-  <version>0.1.0</version>
+  <version>0.2.0</version>
   <description>
 
      op3_gui_demo
@@ -24,6 +24,7 @@
   <build_depend>eigen_conversions</build_depend>
   <build_depend>robotis_controller_msgs</build_depend>
   <build_depend>op3_walking_module_msgs</build_depend>
+  <build_depend>op3_online_walking_module_msgs</build_depend>
   <build_depend>humanoid_nav_msgs</build_depend>
   <run_depend>qt_build</run_depend>
   <run_depend>libqt4-dev</run_depend>
@@ -38,5 +39,6 @@
   <run_depend>eigen_conversions</run_depend>
   <run_depend>robotis_controller_msgs</run_depend>
   <run_depend>op3_walking_module_msgs</run_depend>
+  <run_depend>op3_online_walking_module_msgs</run_depend>
   <run_depend>humanoid_nav_msgs</run_depend>
 </package>
