+++ op3_gui_demo/ui/preview_walking_form.ui
@@ -303,7 +303,7 @@
       <item row="9" column="1">
        <widget class="QPushButton" name="button_send_body_offset">
         <property name="text">
-         <string>Send Offset</string>
+         <string>Set Body Offset</string>
         </property>
        </widget>
       </item>
@@ -462,7 +462,7 @@
       <item row="11" column="1">
        <widget class="QPushButton" name="button_send_foot_distance">
         <property name="text">
-         <string>Send Foot Distance</string>
+         <string>Set Foot Distance</string>
         </property>
        </widget>
       </item>
@@ -584,7 +584,7 @@
       <item row="0" column="0">
        <widget class="QLabel" name="label_37">
         <property name="text">
-         <string>Step TIme</string>
+         <string>Step Time</string>
         </property>
        </widget>
       </item>
